Redesign run button
