+++ models/test-management/src/plugin.ts
@@ -37,7 +37,7 @@ export default mergeIds(testManagementId, testManganement, {
     TestRunPresenter: '' as AnyComponent,
     EditTestRun: '' as AnyComponent,
     TestSuiteRefPresenter: '' as AnyComponent,
-    RunSelectedTestsButton: '' as AnyComponent,
+    RunButton: '' as AnyComponent,
     TestResultPresenter: '' as AnyComponent,
     EditTestResult: '' as AnyComponent,
     TestResultFooter: '' as AnyComponent
